+++ backend/tests/grades_test.go
@@ -49,7 +49,7 @@ func TestGETGrades(t *testing.T) {
 
 	t.Run("returns all grades after some insertions", func(t *testing.T) {
 		defer cleanStore()
-		grade := services.Grade{Subject: "subject_test", Type: "type_test", Value: 10, Student: "student_test"}
+		grade := services.Grade{Subject: "subject_test", Type: "type_test", Value: "10", Student: "student_test"}
 
 		request := utils.NewGetGradesRequest()
 		response := httptest.NewRecorder()
@@ -60,8 +60,8 @@ func TestGETGrades(t *testing.T) {
 		router.ServeHTTP(response, request)
 
 		want := []services.Grade{
-			{ID: "0", Subject: "subject_test", Type: "type_test", Value: 10, Student: "student_test"},
-			{ID: "1", Subject: "subject_test", Type: "type_test", Value: 10, Student: "student_test"},
+			{ID: "0", Subject: "subject_test", Type: "type_test", Value: "10", Student: "student_test"},
+			{ID: "1", Subject: "subject_test", Type: "type_test", Value: "10", Student: "student_test"},
 		}
 		wantJSON, _ := json.Marshal(want)
 		utils.AssertResponseBody(t, response.Body.String(), utils.ResultMessageAndData(utils.AllGrades, string(wantJSON[:])))
@@ -87,7 +87,7 @@ func TestGETGrade(t *testing.T) {
 		request := utils.NewGetGradeRequest("0")
 		response := httptest.NewRecorder()
 
-		grade := services.Grade{ID: "0", Subject: "subject_test", Type: "type_test", Value: 10, Student: "student_test"}
+		grade := services.Grade{ID: "0", Subject: "subject_test", Type: "type_test", Value: "10", Student: "student_test"}
 		router.ServeHTTP(httptest.NewRecorder(), utils.NewPostGradeRequest(grade))
 
 		router.ServeHTTP(response, request)
@@ -103,7 +103,7 @@ func TestPOSTGrade(t *testing.T) {
 
 	t.Run("insert a grade with invalid parameters", func(t *testing.T) {
 		defer cleanStore()
-		grade := services.InvalidGrade{ID: "0", Subject: map[string]string{}, Type: "type_test", Value: 10, Student: "student_test"}
+		grade := services.InvalidGrade{ID: "0", Subject: map[string]string{}, Type: "type_test", Value: "10", Student: "student_test"}
 		request := utils.NewPostInvalidGradeRequest(grade)
 		response := httptest.NewRecorder()
 
@@ -115,7 +115,7 @@ func TestPOSTGrade(t *testing.T) {
 
 	t.Run("insert a grade with missing parameters", func(t *testing.T) {
 		defer cleanStore()
-		grade := services.Grade{ID: "0", Type: "type_test", Value: 10, Student: "student_test"}
+		grade := services.Grade{ID: "0", Type: "type_test", Value: "10", Student: "student_test"}
 		request := utils.NewPostGradeRequest(grade)
 		response := httptest.NewRecorder()
 
@@ -127,7 +127,7 @@ func TestPOSTGrade(t *testing.T) {
 
 	t.Run("insert a grade with value less than 0", func(t *testing.T) {
 		defer cleanStore()
-		grade := services.Grade{ID: "0", Subject: "subject_test", Type: "type_test", Value: -10, Student: "student_test"}
+		grade := services.Grade{ID: "0", Subject: "subject_test", Type: "type_test", Value: "-10", Student: "student_test"}
 		request := utils.NewPostGradeRequest(grade)
 		response := httptest.NewRecorder()
 
@@ -139,7 +139,7 @@ func TestPOSTGrade(t *testing.T) {
 
 	t.Run("insert a grade", func(t *testing.T) {
 		defer cleanStore()
-		grade := services.Grade{ID: "0", Subject: "subject_test", Type: "type_test", Value: 10, Student: "student_test"}
+		grade := services.Grade{ID: "0", Subject: "subject_test", Type: "type_test", Value: "10", Student: "student_test"}
 		request := utils.NewPostGradeRequest(grade)
 		response := httptest.NewRecorder()
 
@@ -155,7 +155,7 @@ func TestPUTGrade(t *testing.T) {
 
 	t.Run("edit a grade with invalid parameters", func(t *testing.T) {
 		defer cleanStore()
-		grade := services.InvalidGrade{ID: "0", Subject: map[string]string{}, Type: "type_test", Value: 10, Student: "student_test"}
+		grade := services.InvalidGrade{ID: "0", Subject: map[string]string{}, Type: "type_test", Value: "10", Student: "student_test"}
 		request := utils.NewEditInvalidGradeRequest("0", grade)
 		response := httptest.NewRecorder()
 
@@ -167,7 +167,7 @@ func TestPUTGrade(t *testing.T) {
 
 	t.Run("edit a grade with missing parameters", func(t *testing.T) {
 		defer cleanStore()
-		grade := services.Grade{ID: "0", Type: "type_test", Value: 10, Student: "student_test"}
+		grade := services.Grade{ID: "0", Type: "type_test", Value: "10", Student: "student_test"}
 		request := utils.NewEditGradeRequest("0", grade)
 		response := httptest.NewRecorder()
 
@@ -179,7 +179,7 @@ func TestPUTGrade(t *testing.T) {
 
 	t.Run("edit a grade with value less than 0", func(t *testing.T) {
 		defer cleanStore()
-		grade := services.Grade{ID: "0", Subject: "subject_test", Type: "type_test", Value: -10, Student: "student_test"}
+		grade := services.Grade{ID: "0", Subject: "subject_test", Type: "type_test", Value: "-10", Student: "student_test"}
 		request := utils.NewEditGradeRequest("0", grade)
 		response := httptest.NewRecorder()
 
@@ -191,7 +191,7 @@ func TestPUTGrade(t *testing.T) {
 
 	t.Run("edit a non existent grade", func(t *testing.T) {
 		defer cleanStore()
-		grade := services.Grade{ID: "5", Subject: "subject_test", Type: "type_test", Value: 10, Student: "student_test"}
+		grade := services.Grade{ID: "5", Subject: "subject_test", Type: "type_test", Value: "10", Student: "student_test"}
 		request := utils.NewEditGradeRequest("5", grade)
 		response := httptest.NewRecorder()
 
@@ -203,7 +203,7 @@ func TestPUTGrade(t *testing.T) {
 
 	t.Run("edit a grade", func(t *testing.T) {
 		defer cleanStore()
-		grade := services.Grade{ID: "0", Subject: "subject_test", Type: "type_test", Value: 10, Student: "student_test"}
+		grade := services.Grade{ID: "0", Subject: "subject_test", Type: "type_test", Value: "10", Student: "student_test"}
 
 		request := utils.NewEditGradeRequest("0", grade)
 		response := httptest.NewRecorder()
@@ -233,7 +233,7 @@ func TestDELETEGrade(t *testing.T) {
 
 	t.Run("delete a grade successfully", func(t *testing.T) {
 		defer cleanStore()
-		grade := services.Grade{ID: "0", Subject: "subject_test", Type: "type_test", Value: 10, Student: "student_test"}
+		grade := services.Grade{ID: "0", Subject: "subject_test", Type: "type_test", Value: "10", Student: "student_test"}
 
 		request := utils.NewDeleteGradeRequest("0")
 		response := httptest.NewRecorder()
@@ -253,10 +253,10 @@ func TestGETGradesByStudent(t *testing.T) {
 
 	t.Run("get all grades by a student", func(t *testing.T) {
 		defer cleanStore()
-		grade1 := services.Grade{ID: "0", Subject: "subject1", Type: "type1", Value: 10, Student: "student1"}
-		grade2 := services.Grade{ID: "1", Subject: "subject2", Type: "type2", Value: 10, Student: "student2"}
-		grade3 := services.Grade{ID: "2", Subject: "subject1", Type: "type1", Value: 10, Student: "student1"}
-		grade4 := services.Grade{ID: "3", Subject: "subject2", Type: "type2", Value: 10, Student: "student2"}
+		grade1 := services.Grade{ID: "0", Subject: "subject1", Type: "type1", Value: "10", Student: "student1"}
+		grade2 := services.Grade{ID: "1", Subject: "subject2", Type: "type2", Value: "10", Student: "student2"}
+		grade3 := services.Grade{ID: "2", Subject: "subject1", Type: "type1", Value: "10", Student: "student1"}
+		grade4 := services.Grade{ID: "3", Subject: "subject2", Type: "type2", Value: "10", Student: "student2"}
 
 		router.ServeHTTP(httptest.NewRecorder(), utils.NewPostGradeRequest(grade1))
 		router.ServeHTTP(httptest.NewRecorder(), utils.NewPostGradeRequest(grade2))
@@ -269,8 +269,8 @@ func TestGETGradesByStudent(t *testing.T) {
 		router.ServeHTTP(response, request)
 
 		want := []services.Grade{
-			{ID: "0", Subject: "subject1", Type: "type1", Value: 10, Student: "student1"},
-			{ID: "2", Subject: "subject1", Type: "type1", Value: 10, Student: "student1"},
+			{ID: "0", Subject: "subject1", Type: "type1", Value: "10", Student: "student1"},
+			{ID: "2", Subject: "subject1", Type: "type1", Value: "10", Student: "student1"},
 		}
 		wantJSON, _ := json.Marshal(want)
 		utils.AssertResponseBody(t, response.Body.String(), utils.ResultMessageAndData(utils.GradesByStudent, string(wantJSON[:])))