+++ frontend/tests/browser.py
@@ -1,0 +1,28 @@
+from selenium import webdriver
+from selenium.webdriver.chrome.options import Options
+
+class Browser(object):
+    options = Options()
+    options.add_argument("--headless")
+    options.add_argument("--no-sandbox")
+    options.add_argument("start-maximized")
+    options.add_argument("disable-infobars")
+    options.add_argument("--disable-extensions")
+
+    # Inicia o browser chrome, mas pode ser feito com outros como Firefox, Safari e IE
+    driver = webdriver.Chrome(chrome_options=options)
+    # Define o tempo máximo para carregamento da página
+    driver.set_page_load_timeout(30)
+    # Maximiza a janela do browser ao ser iniciado
+    driver.maximize_window()
+
+    # Fecha o browser
+    def browser_quit(self):
+        self.driver.quit()
+
+    # Limpa o browser
+    def browser_clear(self):
+        self.driver.delete_all_cookies()
+        self.driver.execute_script('window.localStorage.clear()')
+        self.driver.execute_script('window.sessionStorage.clear()')
+        self.driver.refresh()